Move the enums.txt generator to a new file, experimental.py
